[didelphis-common-39] Update Changelog for 0.3.0  - updated parent version to 0.3.2  - moved `GraphUtils` class
